Moved lib to shared
