+++ src/app/user/HeroSection/HeroSection.tsx
@@ -8,8 +8,8 @@ import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { HeroItem, HeroItemProps } from './HeroItem'
 import { HeroButton } from './HeroButton'
-import { ConnectWorkflow } from '@/lib/walletConnection'
-import { ConnectButtonComponentProps } from '@/lib/walletConnection/types'
+import { ConnectWorkflow } from '@/shared/walletConnection'
+import { ConnectButtonComponentProps } from '@/shared/walletConnection/types'
 
 export const HeroSection = () => {
   const modal = useModal()
