+++ src/components/MainContainer/MainContainer.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useAlertContext } from '@/app/providers'
 import { Footer } from '@/components/Footer'
-import { TopPageHeader } from '@/lib/walletConnection/components/topPageHeader/TopPageHeader'
+import { TopPageHeader } from '@/shared/walletConnection/components/topPageHeader/TopPageHeader'
 import { StatefulSidebar } from '@/components/MainContainer/StatefulSidebar'
 import { FC, ReactNode, Suspense } from 'react'
 import { Alert } from '../Alert'
