+++ src/shared/walletConnection/components/index.ts
