+++ src/shared/walletConnection/components/topPageHeader/TopPageHeaderLeftSlotStrategy.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { usePathname } from 'next/navigation'
-import { getLeftComponentForRoute } from '@/lib/walletConnection/utils'
+import { getLeftComponentForRoute } from '../../utils'
 
 /**
  * Uses Strategy Algorithm to know when to render the left components which vary from page to page
