+++ src/shared/walletConnection/components/topPageHeader/index.ts
