+++ src/shared/walletConnection/components/topPageHeader/pages/UserTopLeftComponent.tsx
