+++ src/shared/walletConnection/components/topPageHeader/pages/index.ts
