+++ src/shared/walletConnection/connection/UserConnectionManager.tsx
