+++ src/shared/walletConnection/connection/index.ts
