+++ src/shared/walletConnection/constants.tsx
