+++ src/shared/walletConnection/index.ts
