+++ src/shared/walletConnection/types.ts
