+++ src/shared/walletConnection/utils.ts
